repos/score-technologies/simlearn/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "simlearn"
version = "0.1.0"
description = "Deep learning models for game outcome prediction"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.0.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
]

[tool.setuptools.packages.find]
include = ["simlearn*"] 